+++ atlas/callbacks.py
@@ -128,27 +128,27 @@ def on_insert_sites(items):
     log.debug(items)
     for item in items:
         log.debug(item)
-        if not item.get('sid'):
+        if 'sid' not in item:
             item['sid'] = 'p1' + sha1(utilities.randomstring()).hexdigest()[0:10]
-        if not item.get('path'):
+        if 'path' not in item:
             item['path'] = item['sid']
-        if not item.get('update_group'):
+        if 'update_group' not in item:
             item['update_group'] = random.randint(0, 2)
         # Add default core and profile if not set.
         # The 'get' method checks if the key exists.
-        if item.get('code'):
-            if not item['code'].get('core'):
-                item['code']['core'] = utilities.get_current_code(
-                    name=DEFAULT_CORE, code_type='core')
-            if not item['code'].get('profile'):
-                item['code']['profile'] = utilities.get_current_code(
-                    name=DEFAULT_PROFILE, code_type='profile')
+        if 'code' in item:
+            if 'core' not in item['code']:
+                item['code']['core'] = ObjectId(utilities.get_current_code(
+                    name=DEFAULT_CORE, code_type='core'))
+            if 'profile' not in item['code']:
+                item['code']['profile'] = ObjectId(utilities.get_current_code(
+                    name=DEFAULT_PROFILE, code_type='profile'))
         else:
             item['code'] = {}
-            item['code']['core'] = utilities.get_current_code(
-                name=DEFAULT_CORE, code_type='core')
-            item['code']['profile'] = utilities.get_current_code(
-                name=DEFAULT_PROFILE, code_type='profile')
+            item['code']['core'] = ObjectId(utilities.get_current_code(
+                name=DEFAULT_CORE, code_type='core'))
+            item['code']['profile'] = ObjectId(utilities.get_current_code(
+                name=DEFAULT_PROFILE, code_type='profile'))
         date_json = '{{"created":"{0} GMT"}}'.format(item['_created'])
         item['dates'] = json.loads(date_json)
 
@@ -234,7 +234,15 @@ def on_delete_item_code(item):
     :param item:
     """
     log.debug('code | on delete | item - %s', item)
-    tasks.code_remove.delay(item)
+    other_static_assets = False
+    if item['meta']['code_type'] == 'static':
+        query = 'where={{"meta.name":"{0}","meta.code_type":"static","_id":{{"$ne":"{1}"}}}}'.format(
+            item['meta']['name'], item['_id'])
+        code = utilities.get_eve('code', query)
+        if code['_meta']['total'] != 0:
+            other_static_assets = True
+    log.info('code | on delete | other static assets - %s', other_static_assets)
+    tasks.code_remove.delay(item, other_static_assets)
 
 
 def on_delete_item_backup(item):